Remove crew from local storage
Add a button to remove a crew previously stored in local storage.
Because localStorage is not part of React and React itself can't
react (lol) on changes, I'd reload the page after an item is
removed from localStorage. Maybe there is a better solution to
that problem without wrapping localStorage into a React state
variable?